File, a dict

+++ googl/file.py
@@ -4,7 +4,7 @@ import magic
 from functools import partial
 from typing import Self
 from pathlib import Path, PurePath
-from dataclasses import dataclass, field, asdict
+from dataclasses import dataclass, field
 
 from googleapiclient.http import MediaFileUpload, MediaIoBaseDownload
 
@@ -15,35 +15,28 @@ from dictionary import gets
 from relation import Relation
 from . import Query, api, request
 
-FIELDS = {'name', 'mimeType', 'id', 'parents'}
+
 FOLDER_MIMETYPE = 'application/vnd.google-apps.folder'
 
 files = api.drive.files()
 
 def _equal_contents(p: Path, drive: File) -> bool:
     return Relation(list(p.iterdir()), \
-        File(parents=[drive.id]).list(), equal).bijection()
+        File(parents=[drive.id]).match(), equal).bijection()
 
 
 def equal(p: Path, drive: File) -> bool:
     return p.name == drive.name \
         and (p.is_dir() == (drive.mimeType == FOLDER_MIMETYPE)) \
         and _equal_contents(p, drive) if p.is_dir() else \
-        file.eq_contents(p, [drive], lambda f: File.content(f.id))
+        file.eq_contents(p, [drive], lambda f: f.content())
 
 
-@dataclass
-class File:
-    mimeType: str | None = field(default=None)
-    parents: list[str] | None = field(default=None)
-    owners: dict | None = field(default=None)
-    id: str = field(default=None)
-    name: str = field(default=None)
-
+class File(dict):
     @staticmethod
     def sync(local: Path, folder: PurePath=PurePath('/')) -> File:
         return File._sync(local, l.one(File.at(folder), File.create,
-                                       File.delete).id)
+                                       File.delete))
     
     @staticmethod
     def _sync(local: Path, parent: str='root') -> File:
@@ -51,14 +44,14 @@ class File:
         if local.is_file():
             f.mimeType = magic.from_file(local, mime=True)
             with open(local, 'rb') as c:
-                return l.one(sorted(f.list(), key=
-                        lambda f: file.eq([c.read(), File.content(f.id)]),
+                return l.one(sorted(f.match(), key=
+                        lambda f: file.eq([c.read(), f.content()]),
                         reverse=True),
                     partial(f.create, MediaFileUpload(str(local))),
                     File.delete)
         if local.is_dir():
             f.mimeType = FOLDER_MIMETYPE
-            f = l.one(f.list(), f.create, f.delete)
+            f = l.one(f.match(), f.create, f.delete)
             for s in local.iterdir(): File._sync(s, f.id)
         return f
 
@@ -67,42 +60,37 @@ class File:
         print(vars(self))
         if media_body: uploadType = 'multipart'
         self.__dict__ |= api.request(files.create(uploadType=uploadType,
-            body=asdict(self), media_body=media_body))
+            body=self, media_body=media_body))
         return self
 
     @staticmethod
     def at(drive=PurePath('/')) -> list[File] | None:
         drive = PurePath(drive)
         if path.top_level(drive): return [File(id='root')]
-        if parents := [f.id for f in File.at(drive.parent)]:
-            return File(parents=parents, name=drive.name).list()
+        if parents := [f['id'] for f in File.at(drive.parent)]:
+            return File(parents=parents, name=drive.name).match()
         return None
 
     def delete(self):
-        self.__dict__ |= request(files.delete(fileId=self.id))
+        request(files.delete(fileId=self['id']))
         
-    def list(self, pattern=None) -> list[File]:
-        return [File(**d) for d in File._list(Query.build(asdict(self),
-            pattern))]
+    def match(self, attrs: set[str]=None, pattern=None) -> list[File]:
+        if not attrs: attrs = set(self.keys()) | {'id'}
+        return File.list(Query.build(self, pattern), attrs)
 
     @staticmethod
-    def _list(q: Query, pageToken: str=None) -> list[dict]:
-        ds, t = gets(File._page(q, pageToken), {'files': [],
-                                                'nextPageToken': []})
-        return ds + (t and File._list(q, t))
+    def list(q: Query, attrs: set[str]=None, pageToken: str=None) \
+            -> list[File]:
+        r, pageToken = gets(request(files.list(q=q, fields= \
+            f"files({','.join(attrs)}),nextPageToken", pageToken=pageToken)),
+            {'files': [], 'nextPageToken': []})
+        return r + (pageToken and File._list(q, attrs, pageToken))
 
-    @staticmethod
-    def _page(q: Query, pageToken: str) -> list[dict]:
-        return request(files.list(q=q, fields=("files(" \
-            + ','.join(File.__dataclass_fields__.keys()) + "),nextPageToken"),
-            pageToken=pageToken))
-
-    @staticmethod
-    def content(id: str) -> bytes:
-        if int(api.request(files.get(fileId=id, fields='size')) \
+    def content(self) -> bytes | None:
+        if int(api.request(files.get(fileId=self.id, fields='size')) \
                .get('size'), 0) == 0: return None
         fh = io.BytesIO()
-        request = files.get_media(fileId=id)
+        request = files.get_media(fileId=self.id)
         downloader = MediaIoBaseDownload(fh, request)
         while not downloader.next_chunk()[1]: continue
         return fh.getvalue()
@@ -114,12 +102,10 @@ class File:
     @staticmethod
     def delete_by_prefix(fs: list[File], prefix: str):
         for f in fs:
-            if f.name.startswith(prefix):
-                f.delete()
+            if f.name.startswith(prefix): f.delete()
 
-    @staticmethod
-    def permissions(id):
-        permissions = File.service.permissions().list(fileId=id).execute()
+    def permissions(self):
+        permissions = File.service.permissions().list(fileId=self.id).execute()
         for permission in permissions.get('permissions', []):
             print(f"ID: {permission['id']}, Type: {permission['type']}, \
                   Role: {permission['role']}")
